structural: fix SonarQube code smells in main source
Fix comment-swallowed braces in record compact constructors and empty
method bodies where { // comment } caused the closing brace to be
consumed by the single-line comment, producing unparseable Java.

Also apply S7467 (unnamed patterns), S6204 (Stream.toList), S1170
(static final constants), S1068 (unused fields), S6205 (redundant
blocks), S1905 (unnecessary casts), S1612 (method references), S6878
(record patterns), S1130 (unused throws), S6353 (use \d in regex),
S1118 (private utility constructor), and S2094 (empty class removal).

Revert three incorrect S7467 unnamed pattern conversions where the
caught exception variable was still referenced in the catch body.

Revert two incorrect S1170 static field conversions on MetadataProvider
implementations where @Getter generates instance methods to satisfy the
interface contract.

+++ src/main/java/com/streamarr/server/services/streaming/HlsPlaylistService.java
@@ -77,7 +77,7 @@ public class HlsPlaylistService {
     }
 
     for (int i = 0; i < segmentCount; i++) {
-      var remainingMs = remainingDurationMs - ((long) i * segmentDurationMs);
+      var remainingMs = remainingDurationMs - (i * segmentDurationMs);
       var durationMs = Math.min(segmentDurationMs, remainingMs);
       sb.append("#EXTINF:").append(String.format("%.6f", durationMs / 1000.0)).append(",\n");
       sb.append("segment").append(i).append(extension).append("\n");
